utils restructure - #124
