+++ pysat/utils/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""
+pysat.utils - utilities for running pysat
+=========================================
+
+pysat.utils contains a number of functions used
+throughout the pysat package.  This includes conversion
+of formats, loading of files, and user-supplied info
+for the pysat data directory structure.
+"""
+
+from . import coords, misc, stats, time
